+++ perses/tests/testsystems.py
@@ -342,18 +342,6 @@ class AlkanesTestSystem(PersesTestSystem):
         self.sams_samplers = sams_samplers
         self.designer = designer
 
-def show_topology(topology):
-    output = ""
-    for atom in topology.atoms():
-        output += "%8d %5s %5s %3s: bonds " % (atom.index, atom.name, atom.residue.id, atom.residue.name)
-        for bond in atom.residue.bonds():
-            if bond[0] == atom:
-                output += " %8d" % bond[1].index
-            if bond[1] == atom:
-                output += " %8d" % bond[0].index
-        output += '\n'
-    print(output)
-
 def check_topologies(testsystem):
     """
     Check that all SystemGenerators can build systems for their corresponding Topology objects.
@@ -366,6 +354,7 @@ def check_topologies(testsystem):
             msg = str(e)
             msg += '\n'
             msg += "topology for environment '%s' cannot be built into a system" % environment
+            from perses.tests.utils import show_topology
             show_topology(topology)
             raise Exception(msg)
 
@@ -375,7 +364,6 @@ def test_AlanineDipeptideTestSystem():
     """
     from perses.tests.testsystems import AlanineDipeptideTestSystem
     testsystem = AlanineDipeptideTestSystem()
-    show_topology(testsystem.topologies['explicit'])
     # Check topologies
     check_topologies(testsystem)
     # Build a system
