+++ perses/tests/utils.py
@@ -39,6 +39,18 @@ beta = 1.0/kT
 
 # TODO: Move some of these utility routines to openmoltools.
 
+def show_topology(topology):
+    output = ""
+    for atom in topology.atoms():
+        output += "%8d %5s %5s %3s: bonds " % (atom.index, atom.name, atom.residue.id, atom.residue.name)
+        for bond in atom.residue.bonds():
+            if bond[0] == atom:
+                output += " %8d" % bond[1].index
+            if bond[1] == atom:
+                output += " %8d" % bond[0].index
+        output += '\n'
+    print(output)
+    
 def extractPositionsFromOEMOL(molecule):
     positions = unit.Quantity(np.zeros([molecule.NumAtoms(), 3], np.float32), unit.angstroms)
     coords = molecule.GetCoords()
